+++ src/core/Cline.ts
@@ -3489,7 +3489,8 @@ export class Cline {
 		details += "\n\n# Current Mode"
 		if (this.chatSettings.mode === "plan") {
 			details += "\nPLAN MODE"
-			details += '\nSee "## What is PLAN MODE?" above for more information about what to do in this mode.'
+			details +=
+				"\nSee \"## What is PLAN MODE?\" above for more information about what to do in this mode. If you haven't done so already, it's a good idea to start by asking a question."
 			details +=
 				'\n(Remember: You now only have access to the plan_mode_response tool. If it seems the user wants you to use tools only available in ACT MODE, you should ask the user to "toggle to Act mode" - they will have to manually do this themselves with the Plan/Act toggle button below. You do not have the ability to switch to ACT MODE yourself, and must wait for the user to do it themselves once they are satisfied with the plan.)'
 		} else {